backend/requirements.txt
flask
flask-cors
dotenv
sqlalchemy
pdfminer.six
google-auth
google-auth-oauthlib
google-auth-httplib2
openai
